#2899: implement layout to prevent circular imports
